fix(listener): keep internal turn failures masked in terminal errors
getConsumerLoopErrorMessage passes unclassified errors through so
channels see the same detail interactive consumers get from loop_error.
That is right for stream- and exception-derived errors, but it also
unmasked internally fabricated messages like "requires_approval stop
returned no approvals". Route those through the generic-fallback
variant again.

Also split the gateway turn-lifecycle tests into their own file: the
merge with main pushed gateway-core.test.ts past the 1,000-line limit,
so its fakes now live in gateway-test-support.ts.
