+++ src/channels/gateway-core.test.ts
@@ -1,286 +1,22 @@
 import { expect, test } from "bun:test";
-import type { MessageCreate } from "@letta-ai/letta-client/resources/agents/agents";
 import type {
   ApprovalResponseBody,
   ExternalToolCallRequestMessage,
-  ExternalToolCallResult,
-  ExternalToolDefinitionPayload,
-  InputAcceptedResponseMessage,
-  RuntimeScope,
   RuntimeStartResponseMessage,
-  StreamDeltaMessage,
   WsProtocolMessage,
 } from "@/types/app-server-protocol";
-import type { QueueUpdateMessage } from "@/types/protocol_v2";
+import { ChannelGateway } from "./gateway-core";
 import {
-  ChannelGateway,
-  type ChannelGatewayClient,
-  type ChannelGatewayDelivery,
-  type ChannelGatewayHooks,
-} from "./gateway-core";
-import type {
-  ChannelControlRequestEvent,
-  ChannelTurnLifecycleEvent,
-  ChannelTurnProgressEvent,
-  ChannelTurnSource,
-} from "./types";
-
-// ── Fake client ──────────────────────────────────────────────────
-
-interface FakeClientOptions {
-  startResponse?: Partial<RuntimeStartResponseMessage>;
-  inputResponse?: Partial<InputAcceptedResponseMessage>;
-  toolUpdateWait?: Promise<void>;
-}
-
-class FakeClient implements ChannelGatewayClient {
-  private messageListeners: Array<(message: WsProtocolMessage) => void> = [];
-  private externalToolListeners: Array<
-    (request: ExternalToolCallRequestMessage) => unknown
-  > = [];
-  readonly submittedInputs: Array<{
-    runtime: RuntimeScope;
-    payload: unknown;
-  }> = [];
-  readonly startedRuntimes: Array<{
-    agent_id?: string;
-    conversation_id?: string;
-    mode?: string;
-    preserve_skill_sources?: boolean;
-    external_tools?: unknown;
-  }> = [];
-  readonly runtimeToolUpdates: Array<
-    import("@/types/app-server-protocol").RuntimeExternalToolsUpdateGroup
-  > = [];
-  closeCalls = 0;
-  // Mutable input response for per-test control
-  inputResponse: { accepted: boolean; disposition: "started" | "queued" };
-
-  constructor(private readonly options: FakeClientOptions = {}) {
-    this.inputResponse = {
-      accepted: options.inputResponse?.accepted ?? true,
-      disposition: options.inputResponse?.disposition ?? "started",
-    };
-  }
-
-  onMessage(listener: (message: WsProtocolMessage) => void): () => void {
-    this.messageListeners.push(listener);
-    return () => {
-      const idx = this.messageListeners.indexOf(listener);
-      if (idx >= 0) this.messageListeners.splice(idx, 1);
-    };
-  }
-
-  onExternalToolCall(
-    handler: (
-      request: ExternalToolCallRequestMessage,
-    ) => Promise<ExternalToolCallResult> | ExternalToolCallResult,
-  ): () => void {
-    this.externalToolListeners.push(handler);
-    return () => {
-      const idx = this.externalToolListeners.indexOf(handler);
-      if (idx >= 0) this.externalToolListeners.splice(idx, 1);
-    };
-  }
-
-  async submitInput(
-    command: Omit<import("@/types/app-server-protocol").InputCommand, "type">,
-  ): Promise<InputAcceptedResponseMessage> {
-    this.submittedInputs.push({
-      runtime: command.runtime,
-      payload: command.payload,
-    });
-    return {
-      type: "input_accepted",
-      request_id: "test-req",
-      runtime: command.runtime,
-      accepted: this.inputResponse.accepted,
-      disposition: this.inputResponse.disposition,
-      ...(this.options.inputResponse?.error
-        ? { error: this.options.inputResponse.error }
-        : {}),
-    };
-  }
-
-  async runtimeStart(
-    opts: Omit<
-      import("@/types/app-server-protocol").RuntimeStartCommand,
-      "type" | "request_id"
-    > & { request_id?: string },
-  ): Promise<RuntimeStartResponseMessage> {
-    this.startedRuntimes.push(opts);
-    return {
-      type: "runtime_start_response",
-      request_id: opts.request_id ?? "test-req",
-      success: this.options.startResponse?.success ?? true,
-      runtime: this.options.startResponse?.runtime ?? null,
-      agent: this.options.startResponse?.agent ?? null,
-      conversation: this.options.startResponse?.conversation ?? null,
-      created: this.options.startResponse?.created ?? {
-        agent: false,
-        conversation: false,
-      },
-      ...(this.options.startResponse?.error
-        ? { error: this.options.startResponse.error }
-        : {}),
-    };
-  }
-
-  async runtimeExternalToolsUpdate(options: {
-    updates: readonly import("@/types/app-server-protocol").RuntimeExternalToolsUpdateGroup[];
-  }): Promise<
-    import("@/types/app-server-protocol").RuntimeExternalToolsUpdateResponseMessage
-  > {
-    this.runtimeToolUpdates.push(...options.updates);
-    await this.options.toolUpdateWait;
-    return {
-      type: "runtime_external_tools_update_response",
-      request_id: "runtime-external-tools-test",
-      success: true,
-    };
-  }
-
-  close(): void {
-    this.closeCalls += 1;
-  }
-
-  // Test helpers
-  emit(message: WsProtocolMessage): void {
-    for (const listener of this.messageListeners) listener(message);
-  }
-
-  emitExternalToolCall(request: ExternalToolCallRequestMessage): void {
-    for (const listener of this.externalToolListeners) listener(request);
-  }
-}
-
-// ── Helpers ──────────────────────────────────────────────────────
-
-const TEST_RUNTIME: RuntimeScope = {
-  agent_id: "agent-1",
-  conversation_id: "conv-1",
-};
-
-function makeSource(
-  overrides: Partial<ChannelTurnSource> = {},
-): ChannelTurnSource {
-  return {
-    channel: "telegram",
-    chatId: "chat-1",
-    agentId: "agent-1",
-    conversationId: "conv-1",
-    ...overrides,
-  };
-}
-
-function makeDelivery(
-  overrides: Partial<ChannelGatewayDelivery> = {},
-): ChannelGatewayDelivery {
-  return {
-    runtime: TEST_RUNTIME,
-    content: "Hello" as MessageCreate["content"],
-    sources: [makeSource()],
-    clientMessageId: "cm-test-1",
-    ...overrides,
-  };
-}
-
-interface HookCollector {
-  hooks: ChannelGatewayHooks;
-  lifecycleEvents: ChannelTurnLifecycleEvent[];
-  progressEvents: ChannelTurnProgressEvent[];
-  controlRequestEvents: ChannelControlRequestEvent[];
-  externalToolResults: ExternalToolCallResult[];
-}
-
-function makeHooks(
-  overrides: Partial<ChannelGatewayHooks> = {},
-): HookCollector {
-  const lifecycleEvents: ChannelTurnLifecycleEvent[] = [];
-  const progressEvents: ChannelTurnProgressEvent[] = [];
-  const controlRequestEvents: ChannelControlRequestEvent[] = [];
-  const externalToolResults: ExternalToolCallResult[] = [];
-
-  const hooks: ChannelGatewayHooks = {
-    buildExternalTool: async () =>
-      ({
-        name: "MessageChannel",
-        description: "Send a message through a channel",
-        parameters: {},
-      }) satisfies ExternalToolDefinitionPayload,
-    executeExternalTool: async (_request) => {
-      const result: ExternalToolCallResult = {
-        content: [{ type: "text", text: "ok" }],
-      };
-      externalToolResults.push(result);
-      return result;
-    },
-    onLifecycle: (event) => {
-      lifecycleEvents.push(event);
-    },
-    onProgress: (event) => {
-      progressEvents.push(event);
-    },
-    onControlRequest: (event) => {
-      controlRequestEvents.push(event);
-    },
-    ...overrides,
-  };
-
-  return {
-    hooks,
-    lifecycleEvents,
-    progressEvents,
-    controlRequestEvents,
-    externalToolResults,
-  };
-}
-
-function makeStreamDelta(
-  delta: Record<string, unknown>,
-  runtime: RuntimeScope = TEST_RUNTIME,
-): StreamDeltaMessage {
-  return {
-    type: "stream_delta",
-    runtime,
-    event_seq: 0,
-    emitted_at: new Date().toISOString(),
-    idempotency_key: "key-1",
-    delta: delta as unknown as StreamDeltaMessage["delta"],
-  };
-}
-
-function makeQueueUpdate(
-  queue: Array<{ client_message_id: string }>,
-  runtime: RuntimeScope = TEST_RUNTIME,
-): QueueUpdateMessage {
-  return {
-    type: "update_queue",
-    runtime,
-    event_seq: 0,
-    emitted_at: new Date().toISOString(),
-    idempotency_key: "key-1",
-    queue: queue as unknown as QueueUpdateMessage["queue"],
-  };
-}
-
-function makeTurnFinished(
-  stopReason: string,
-  runtime: RuntimeScope = TEST_RUNTIME,
-  extra: { runId?: string; error?: string } = {},
-): WsProtocolMessage {
-  return {
-    type: "turn_finished",
-    runtime,
-    event_seq: 0,
-    emitted_at: new Date().toISOString(),
-    idempotency_key: "key-1",
-    turn_id: "turn-1",
-    stop_reason: stopReason,
-    ...(extra.runId ? { run_id: extra.runId } : {}),
-    ...(extra.error ? { error: extra.error } : {}),
-  } as unknown as WsProtocolMessage;
-}
+  FakeClient,
+  makeDelivery,
+  makeHooks,
+  makeQueueUpdate,
+  makeSource,
+  makeStreamDelta,
+  makeTurnFinished,
+  TEST_RUNTIME,
+} from "./gateway-test-support";
+import type { ChannelTurnSource } from "./types";
 
 // ── Tests ────────────────────────────────────────────────────────
 
@@ -854,178 +590,6 @@ test("control request with multiple sources does not dispatch", async () => {
 
   expect(controlRequestEvents).toHaveLength(0);
 
-  gateway.close();
-});
-
-test("stream stop reason waits for turn_finished error detail", async () => {
-  const client = new FakeClient();
-  const { hooks, lifecycleEvents, progressEvents } = makeHooks();
-  const gateway = new ChannelGateway(client, hooks);
-
-  await gateway.submit(makeDelivery({ clientMessageId: "cm-stream" }));
-
-  client.emit(
-    makeStreamDelta({
-      message_type: "reasoning_message",
-      run_id: "run-1",
-    }),
-  );
-  expect(progressEvents.length).toBeGreaterThan(0);
-
-  client.emit(
-    makeStreamDelta({
-      message_type: "stop_reason",
-      stop_reason: "insufficient_credits",
-      run_id: "run-1",
-    }),
-  );
-  await Bun.sleep(0);
-  expect(
-    lifecycleEvents.filter((event) => event.type === "finished"),
-  ).toHaveLength(0);
-
-  client.emit(
-    makeTurnFinished("insufficient_credits", TEST_RUNTIME, {
-      runId: "run-1",
-      error: "The usage limit has been reached.",
-    }),
-  );
-
-  const finishedEvents = lifecycleEvents.filter(
-    (event) => event.type === "finished",
-  );
-  expect(finishedEvents).toHaveLength(1);
-  expect(finishedEvents[0]).toMatchObject({
-    outcome: "error",
-    stopReason: "insufficient_credits",
-    runId: "run-1",
-    error: "The usage limit has been reached.",
-  });
-
-  gateway.close();
-});
-
-test("run-level error stop waits for the final turn_finished event", async () => {
-  const client = new FakeClient();
-  const { hooks, lifecycleEvents } = makeHooks();
-  const gateway = new ChannelGateway(client, hooks);
-
-  await gateway.submit(makeDelivery({ clientMessageId: "cm-retry" }));
-  client.emit(
-    makeStreamDelta({
-      message_type: "loop_error",
-      message: "temporary provider failure",
-      is_terminal: false,
-      run_id: "run-failed",
-    }),
-  );
-  client.emit(
-    makeStreamDelta({
-      message_type: "stop_reason",
-      stop_reason: "llm_api_error",
-      run_id: "run-failed",
-    }),
-  );
-  expect(lifecycleEvents.filter((event) => event.type === "finished")).toEqual(
-    [],
-  );
-
-  client.emit(
-    makeTurnFinished("end_turn", TEST_RUNTIME, { runId: "run-retry" }),
-  );
-  await Bun.sleep(0);
-
-  const finished = lifecycleEvents.find((event) => event.type === "finished");
-  expect(finished).toMatchObject({
-    outcome: "completed",
-    runId: "run-retry",
-    stopReason: "end_turn",
-  });
-  expect(finished).not.toHaveProperty("error");
-  gateway.close();
-});
-
-test("subagent stop does not finish or overwrite the parent turn", async () => {
-  const client = new FakeClient();
-  const { hooks, lifecycleEvents } = makeHooks();
-  const gateway = new ChannelGateway(client, hooks);
-
-  await gateway.submit(makeDelivery({ clientMessageId: "cm-subagent" }));
-  client.emit({
-    ...makeStreamDelta({
-      message_type: "stop_reason",
-      stop_reason: "insufficient_credits",
-      run_id: "run-subagent",
-    }),
-    subagent_id: "subagent-1",
-  });
-  expect(lifecycleEvents.filter((event) => event.type === "finished")).toEqual(
-    [],
-  );
-
-  client.emit(
-    makeTurnFinished("end_turn", TEST_RUNTIME, { runId: "run-parent" }),
-  );
-  await Bun.sleep(0);
-
-  const finished = lifecycleEvents.find((event) => event.type === "finished");
-  expect(finished).toMatchObject({
-    outcome: "completed",
-    runId: "run-parent",
-    stopReason: "end_turn",
-  });
-  gateway.close();
-});
-
-test("requires_approval stop reason does not finish the turn", async () => {
-  const client = new FakeClient();
-  const { hooks, lifecycleEvents } = makeHooks();
-  const gateway = new ChannelGateway(client, hooks);
-
-  await gateway.submit(makeDelivery({ clientMessageId: "cm-approval" }));
-
-  // Emit a requires_approval stop_reason
-  client.emit(
-    makeStreamDelta({
-      message_type: "stop_reason",
-      stop_reason: "requires_approval",
-      run_id: "run-1",
-    }),
-  );
-
-  // Should NOT trigger finished lifecycle
-  const finishedEvents = lifecycleEvents.filter((e) => e.type === "finished");
-  expect(finishedEvents).toHaveLength(0);
-
-  gateway.close();
-});
-
-test("terminal requires_approval recovery event clears the active turn", async () => {
-  const client = new FakeClient();
-  const { hooks, lifecycleEvents } = makeHooks();
-  const gateway = new ChannelGateway(client, hooks);
-
-  await gateway.submit(makeDelivery({ clientMessageId: "cm-recovery" }));
-  client.emit(
-    makeTurnFinished("requires_approval", TEST_RUNTIME, {
-      runId: "run-recovery",
-      error: "Recovery continuation ended unexpectedly: requires_approval",
-    }),
-  );
-  await Bun.sleep(0);
-
-  const finished = lifecycleEvents.find((event) => event.type === "finished");
-  expect(finished).toMatchObject({
-    outcome: "error",
-    runId: "run-recovery",
-    stopReason: "requires_approval",
-    error: "Recovery continuation ended unexpectedly: requires_approval",
-  });
-
-  await gateway.submit(makeDelivery({ clientMessageId: "cm-next" }));
-  expect(
-    lifecycleEvents.filter((event) => event.type === "processing"),
-  ).toHaveLength(2);
   gateway.close();
 });
